Rename email to hnet/hmail
