+++ lib/std/hnet/hmail/address.go
@@ -1,4 +1,4 @@
-package email
+package hmail
 
 import (
 	"errors"
