+++ lib/std/hnet/hmail/address_test.go
@@ -1,4 +1,4 @@
-package email
+package hmail
 
 import (
 	"testing"
